Collidersフォルダを追加。 Bullet等を移動。Collider等を移動。
