+++ Assets/Scripts/Colliders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88258153f3588b94e9c8f8f72064d6cf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
